fix: lend market loading
