+++ apps/main/src/loan/store/createAppSlice.ts
@@ -2,13 +2,13 @@ import { produce } from 'immer'
 import lodash from 'lodash'
 import type { Config } from 'wagmi'
 import type { StoreApi } from 'zustand'
-import { prefetchMintMarkets } from '@/loan/entities/mint-markets.query'
 import { type State } from '@/loan/store/useStore'
 import { type LlamaApi, Wallet } from '@/loan/types/loan.types'
 import { log } from '@/loan/utils/helpers'
 import { isLoanSlicesEnabled } from '@ui-kit/hooks/useFeatureFlags'
 import { ReleaseChannel } from '@ui-kit/utils'
 import { formatTimeDiff } from '@ui-kit/utils/time.utils'
+import { prefetchMintMarkets } from '../entities/mint-market-names.query'
 
 export type SliceKey = keyof State | ''
 export type StateKey = string
