+++ tests/cypress/support/helpers/llamalend/LlammalendTestCase.tsx
@@ -1,3 +1,4 @@
+import { prefetchMintMarkets } from 'main/src/loan/entities/mint-market-names.query'
 import { useMemo } from 'react'
 import { prefetchLendMarkets } from '@/lend/queries/lend-market-names.query'
 import { CreateLoanForm } from '@/llamalend/features/borrow/components/CreateLoanForm'
@@ -14,7 +15,6 @@ import { UnstakeForm } from '@/llamalend/features/supply/components/UnstakeForm'
 import { WithdrawForm } from '@/llamalend/features/supply/components/WithdrawForm'
 import { getLlamaMarket } from '@/llamalend/llama.utils'
 import { useLoanExists } from '@/llamalend/queries/user'
-import { prefetchMintMarkets } from '@/loan/entities/mint-markets.query'
 import type { IChainId as LlamaChainId } from '@curvefi/llamalend-api/lib/interfaces'
 import { ComponentTestWrapper } from '@cy/support/helpers/ComponentTestWrapper'
 import { fakeCollateralEvents } from '@cy/support/helpers/llamalend/mock-loan-test-data'
